Fix: use correct local loadshedding year|month
Previously, local load shedding (which usually only has a date and time
specified) would be converted to an actual date time by just using the
current year and month.

This caused issues when national load shedding was in a different year
or month than the current month.

This fix forces the local load shedding to have the same month and year
as the start of national load shedding, which should solve this issue.