Remove invalid kareedouw files (#207)
Turns out the official eskom loadshedding schedule for Kareedouw has
nothing inside of it:
eskom.co.za/distribution/wp-content/uploads/2021/06/Kareedouw-1.pdf
which was why eskom-calendar was showing there to be no loadshedding for
kareedouw.

But it's not a problem because Kareedouw's schedule is identical to
eskom-direct-16-odd, so kareedouw can just use that schedule instead.