Fix: Remove events with duration <= 30mins #407
Load shedding events less than 30 minutes long are not actually
implemented in practice, because it is too much hassle for the
municipalities to turn the areas off for just 30 minutes of
load shedding.

This was previously unaddressed because of the edge case where there is
30 minutes of stage X, immediately followed by 2 hours of stage Y. In
this case, there *is* loadshedding for the full 2h30.

This commit does not address that edge case, but support will be added
in the near future.